Trigonometry has been added 2
