+++ example/math_utils_example.dart
@@ -1,8 +1,1 @@
-import 'package:math_utils_n13/math_utils_n13.dart';
 
-void main() {
-  print(add(2, 3)); // 5
-  print(subtract(10, 4)); // 6
-  print(multiply(5, 2)); // 10
-  print(divide(8, 2)); // 4.0
-}
