+++ lib/src/trigonometry.dart
@@ -1,0 +1,2 @@
+double toDegrees(double radians) => radians * (180 / 3.141592653589793);
+double toRadians(double degrees) => degrees * (3.141592653589793 / 180);
